Added last stuff

+++ CustomerAPI/Models/Customer.cs
@@ -13,7 +13,7 @@ namespace CustomerAPI.Models
         public string Phone { get; set; }
         public string BillingAddress { get; set; }
         public string ShippingAddress { get; set; }
-        public int CreditStanding { get; set; }
+        public decimal CreditStanding { get; set; }
 
     }
 }
